#12 [refactor] lottie 띄우는 로직 main으로 이동
