+++ feature/progress/src/main/java/com/sopetit/progress/ProgressPageState.kt
@@ -7,5 +7,4 @@ import com.sopetit.ui.base.PageState
 data class ProgressPageState(
     val memberDailyRoutineList: List<MemberDailyRoutineListModel> = emptyList(),
     val memberChallenge: MemberChallengeModel = MemberChallengeModel(),
-    val isChallengeAchieveShowValid: Boolean = false,
 ) : PageState
