+++ feature/progress/src/main/java/com/sopetit/progress/ProgressViewModel.kt
@@ -11,6 +11,7 @@ import com.sopetit.domain.usecase.memberroutine.DeleteMemberDailyRoutineUseCase
 import com.sopetit.domain.usecase.memberroutine.GetMemberDailyRoutineUseCase
 import com.sopetit.ui.base.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -24,6 +25,8 @@ class ProgressViewModel @Inject constructor(
 ) : BaseViewModel<ProgressPageState>(
     ProgressPageState()
 ) {
+
+    val isChallengeAchieveShowValid = MutableSharedFlow<Boolean>()
 
     init {
         initGetMemberChallenge()
@@ -91,17 +94,9 @@ class ProgressViewModel @Inject constructor(
             achieveMemberChallengeUseCase(request = Unit).collect {
                 resultResponse(it, {
                     initGetMemberChallenge()
-                    updateChallengeAchieve(isValid = true)
                 })
+                isChallengeAchieveShowValid.emit(true)
             }
         }
-    }
-
-    fun updateChallengeAchieve(isValid: Boolean) {
-        updateState(
-            uiState.value.copy(
-                isChallengeAchieveShowValid = isValid
-            )
-        )
     }
 }
